+++ src/components/common/Loading/Loading.js
@@ -1,36 +1,18 @@
 import React, { useEffect, useState } from "react";
 import { View } from "react-native";
 
-import styled from "@emotion/native";
+import NUMBERS from "../../../constants/numbers";
+import {
+  LoadingContainer,
+  LoadingTitle,
+  LoadingImageContainer,
+  LoadingImage,
+} from "./Loading.style"
 
-import NUMBERS from "../../constants/numbers";
-
-const LoadingContainer = styled.View`
-  flex: 1;
-  justify-content: center;
-  align-items: center;
-`;
-
-const LoadingImageContainer = styled.View`
-  flex-direction: row;
-  height: 70px;
-`;
-
-const LoadingImage = styled.Image`
-  width: 70px;
-  height: 70px;
-  padding: 0 20px;
-`;
-
-const LoadingTitle = styled.Text`
-  font-size: 30px;
-  font-family: ${(props) => props.theme.subFont};
-`;
-
-const LoadingScreen = () => {
+const LoadingPage = () => {
   const [footPrintImages, setFootPrintImages] = useState([]);
   const footPrintImage = (
-    <LoadingImage source={require("../../asset/image/loading.png")} />
+    <LoadingImage source={require("../../../asset/image/loading.png")} />
   );
 
   useEffect(() => {
@@ -55,4 +37,4 @@ const LoadingScreen = () => {
   );
 };
 
-export default LoadingScreen;
+export default LoadingPage;
